Changes line_item processing and shipping_methods
Prevents a sellable line_item's price from changing (due to a change in
the sellable's price). This allows sellable prices to change in the
system without causing orders to become unbalanced. When an order is
created, the price of the item at the time of purchase becomes locked
in their order, though it may still be changed manually.

ShippingMethod calls needed a slight tweak for the jdbc-mysql adapter.